+++ package.json
@@ -27,9 +27,9 @@
     "next": "15.3.1",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
- 
     "react-hook-form": "^7.56.1",
     "react-icons": "^5.5.0",
+    "react-player": "^2.16.0",
     "react-redux": "^9.2.0",
     "redux-persist": "^6.0.0",
     "sonner": "^2.0.3",
